# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/classes/{id}": {
      "get": {
        "operationId": "ClassesController_getById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Classes"
        ]
      },
      "patch": {
        "operationId": "ClassesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateClassDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Classes"
        ]
      }
    },
...
    "/api/v1/enrollments": {
      "post": {
        "operationId": "EnrollmentsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEnrollmentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Enrollments"
        ]
      },
      "get": {
        "operationId": "EnrollmentsController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Enrollments"
        ]
      }
    },
    "/api/v1/enrollments/{id}": {
      "delete": {
        "operationId": "EnrollmentsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Enrollments"
        ]
      }
    },
    "/api/v1/attendance-sessions": {
      "post": {
        "operationId": "AttendanceController_createSession",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAttendanceSessionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Attendance"
        ]
      }
    },
    "/api/v1/attendance-sessions/{id}/records": {
      "post": {
        "operationId": "AttendanceController_upsertRecords",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkAttendanceRecordsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Attendance"
        ]
      },
      "get": {
        "operationId": "AttendanceController_listRecords",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Attendance"
        ]
      }
    },
    "/api/v1/students/{id}/attendance": {
      "get": {
        "operationId": "AttendanceController_history",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Attendance"
        ]
      }
    },
...
    "/api/v1/belt-exams": {
      "get": {
        "operationId": "ExamsController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Exams"
        ]
      },
      "post": {
        "operationId": "ExamsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBeltExamDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Exams"
        ]
      }
    },
    "/api/v1/belt-exams/{id}": {
      "get": {
        "operationId": "ExamsController_getById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Exams"
        ]
      },
      "patch": {
        "operationId": "ExamsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBeltExamDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Exams"
        ]
      }
    },
    "/api/v1/belt-exams/{id}/register": {
      "post": {
        "operationId": "ExamsController_register",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterExamDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Exams"
        ]
      }
    },
    "/api/v1/exam-registrations/{id}/result": {
      "post": {
        "operationId": "ExamsController_recordResult",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExamResultDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Exams"
        ]
      }
    },
...
  },
  "info": {
    "title": "VovinamApiNode",
    "description": "Vovinam club management API",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "UpdateClassDto": {
        "type": "object",
        "properties": {}
      },
...
      "CreateEnrollmentDto": {
        "type": "object",
        "properties": {}
      },
      "CreateAttendanceSessionDto": {
        "type": "object",
        "properties": {}
      },
      "BulkAttendanceRecordsDto": {
        "type": "object",
        "properties": {}
      },
...
      "CreateBeltExamDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateBeltExamDto": {
        "type": "object",
        "properties": {}
      },
      "RegisterExamDto": {
        "type": "object",
        "properties": {}
      },
      "ExamResultDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}